download buttons revised to csv only and documents revised to show transcription snippets
